refactor: move proposal handler code into server/core
