+++ x/ecocredit/server/core/proposal_handler_test.go
@@ -1,72 +1,26 @@
-package server
+package core
 
 import (
-	"context"
 	"testing"
 
-	"github.com/cosmos/cosmos-sdk/orm/model/ormdb"
-	"github.com/cosmos/cosmos-sdk/orm/model/ormtable"
-	"github.com/cosmos/cosmos-sdk/orm/testing/ormtest"
-	"github.com/cosmos/cosmos-sdk/store"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/types/module"
-	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
-	"github.com/golang/mock/gomock"
-	"github.com/tendermint/tendermint/libs/log"
-	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
-	dbm "github.com/tendermint/tm-db"
 	"gotest.tools/v3/assert"
 
 	api "github.com/regen-network/regen-ledger/api/regen/ecocredit/v1"
 	"github.com/regen-network/regen-ledger/types/ormutil"
-	"github.com/regen-network/regen-ledger/x/ecocredit"
 	coretypes "github.com/regen-network/regen-ledger/x/ecocredit/core"
-	"github.com/regen-network/regen-ledger/x/ecocredit/mocks"
-	"github.com/regen-network/regen-ledger/x/ecocredit/server/core"
 )
-
-type TestSuite struct {
-	module     module.AppModule
-	keeper     ProposalKeeper
-	stateStore api.StateStore
-	querier    sdk.Querier
-	handler    govtypes.Handler
-	sdkCtx     sdk.Context
-	ctx        context.Context
-}
-
-func setupTest(t *testing.T) TestSuite {
-	s := TestSuite{}
-	db := dbm.NewMemDB()
-	cms := store.NewCommitMultiStore(db)
-	storeKey := sdk.NewKVStoreKey("test")
-	cms.MountStoreWithDB(storeKey, sdk.StoreTypeIAVL, db)
-	assert.NilError(t, cms.LoadLatestVersion())
-	ormCtx := ormtable.WrapContextDefault(ormtest.NewMemoryBackend())
-	s.sdkCtx = sdk.NewContext(cms, tmproto.Header{}, false, log.NewNopLogger()).WithContext(ormCtx)
-	s.ctx = sdk.WrapSDKContext(s.sdkCtx)
-
-	ormDB, err := ormdb.NewModuleDB(&ecocredit.ModuleSchema, ormdb.ModuleDBOptions{})
-	assert.NilError(t, err)
-	s.stateStore, err = api.NewStateStore(ormDB)
-	ctrl := gomock.NewController(t)
-	bankKeeper := mocks.NewMockBankKeeper(ctrl)
-	paramsKeeper := mocks.NewMockParamKeeper(ctrl)
-	s.keeper = core.NewKeeper(s.stateStore, bankKeeper, paramsKeeper)
-	s.handler = NewCreditTypeProposalHandler(s.keeper)
-	return s
-}
 
 func TestCreditTypeProposal_BasicValid(t *testing.T) {
 	t.Parallel()
-	s := setupTest(t)
+	s := setupBase(t)
+	handler := NewCreditTypeProposalHandler(s.k)
 	ct := &coretypes.CreditType{
 		Abbreviation: "BIO",
 		Name:         "biodiversity",
 		Unit:         "meters squared",
 		Precision:    6,
 	}
-	err := s.handler(s.sdkCtx, &coretypes.CreditTypeProposal{
+	err := handler(s.sdkCtx, &coretypes.CreditTypeProposal{
 		Title:       "my credit type",
 		Description: "yay",
 		CreditType:  ct,
@@ -79,14 +33,15 @@ func TestCreditTypeProposal_BasicValid(t *testing.T) {
 
 func TestCreditTypeProposal_InvalidPrecision(t *testing.T) {
 	t.Parallel()
-	s := setupTest(t)
+	s := setupBase(t)
+	handler := NewCreditTypeProposalHandler(s.k)
 	ct := &coretypes.CreditType{
 		Abbreviation: "BIO",
 		Name:         "biodiversity",
 		Unit:         "meters squared",
 		Precision:    3,
 	}
-	err := s.handler(s.sdkCtx, &coretypes.CreditTypeProposal{
+	err := handler(s.sdkCtx, &coretypes.CreditTypeProposal{
 		Title:       "My New Credit Type",
 		Description: "its very cool and awesome",
 		CreditType:  ct,
@@ -96,14 +51,15 @@ func TestCreditTypeProposal_InvalidPrecision(t *testing.T) {
 
 func TestCreditTypeProposal_InvalidAbbreviation(t *testing.T) {
 	t.Parallel()
-	s := setupTest(t)
+	s := setupBase(t)
+	handler := NewCreditTypeProposalHandler(s.k)
 	ct := &coretypes.CreditType{
 		Abbreviation: "biO",
 		Name:         "biodiversity",
 		Unit:         "meters squared",
 		Precision:    6,
 	}
-	err := s.handler(s.sdkCtx, &coretypes.CreditTypeProposal{
+	err := handler(s.sdkCtx, &coretypes.CreditTypeProposal{
 		Title:       "My New Credit Type",
 		Description: "its very cool and awesome",
 		CreditType:  ct,
@@ -113,13 +69,14 @@ func TestCreditTypeProposal_InvalidAbbreviation(t *testing.T) {
 
 func TestCreditTypeProposal_NoName(t *testing.T) {
 	t.Parallel()
-	s := setupTest(t)
+	s := setupBase(t)
+	handler := NewCreditTypeProposalHandler(s.k)
 	ct := &coretypes.CreditType{
 		Abbreviation: "BIO",
 		Unit:         "meters squared",
 		Precision:    6,
 	}
-	err := s.handler(s.sdkCtx, &coretypes.CreditTypeProposal{
+	err := handler(s.sdkCtx, &coretypes.CreditTypeProposal{
 		Title:       "My New Credit Type",
 		Description: "its very cool and awesome",
 		CreditType:  ct,
@@ -129,13 +86,14 @@ func TestCreditTypeProposal_NoName(t *testing.T) {
 
 func TestCreditTypeProposal_NoUnit(t *testing.T) {
 	t.Parallel()
-	s := setupTest(t)
+	s := setupBase(t)
+	handler := NewCreditTypeProposalHandler(s.k)
 	ct := &coretypes.CreditType{
 		Abbreviation: "BIO",
 		Name:         "FooBar",
 		Precision:    6,
 	}
-	err := s.handler(s.sdkCtx, &coretypes.CreditTypeProposal{
+	err := handler(s.sdkCtx, &coretypes.CreditTypeProposal{
 		Title:       "My New Credit Type",
 		Description: "its very cool and awesome",
 		CreditType:  ct,
@@ -145,7 +103,8 @@ func TestCreditTypeProposal_NoUnit(t *testing.T) {
 
 func TestCreditTypeProposal_Duplicate(t *testing.T) {
 	t.Parallel()
-	s := setupTest(t)
+	s := setupBase(t)
+	handler := NewCreditTypeProposalHandler(s.k)
 	ct := &coretypes.CreditType{
 		Abbreviation: "BIO",
 		Name:         "FooBar",
@@ -155,7 +114,7 @@ func TestCreditTypeProposal_Duplicate(t *testing.T) {
 	var pulsarCreditType api.CreditType
 	assert.NilError(t, ormutil.GogoToPulsarSlow(ct, &pulsarCreditType))
 	assert.NilError(t, s.stateStore.CreditTypeTable().Insert(s.ctx, &pulsarCreditType))
-	err := s.handler(s.sdkCtx, &coretypes.CreditTypeProposal{
+	err := handler(s.sdkCtx, &coretypes.CreditTypeProposal{
 		Title:       "My New Credit Type",
 		Description: "its very cool and awesome",
 		CreditType:  ct,
