+++ x/ecocredit/server/expected_keepers.go
@@ -3,16 +3,11 @@ package server
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
-	"github.com/regen-network/regen-ledger/x/ecocredit/core"
+	core2 "github.com/regen-network/regen-ledger/x/ecocredit/server/core"
 )
 
 // Keeper defines the expected interface needed to prune expired buy and sell orders.
 type Keeper interface {
-	ProposalKeeper
+	core2.ProposalKeeper
 	PruneOrders(ctx sdk.Context) error
-}
-
-// ProposalKeeper defines the expected interface for ecocredit module proposals.
-type ProposalKeeper interface {
-	NewCreditType(ctx sdk.Context, ctp *core.CreditTypeProposal) error
 }
